# openapi.yaml
openapi: 3.0.3
info:
  title: Generalized Exam Item Bank & Paper Generation API
  version: 1.0.0
  description: >-
    REST contract for RBAC-secured question management, paper assembly, online examination,
    and analytics services powering the Universal Question Bank Paper System.
servers:
  - url: /api
    description: Development
  - url: /api
    description: Production
security:
  - bearerAuth: []
paths:
...
  /api/exams/{accessCode}/start:
    post:
      summary: Validate access code and start exam session.
      tags: [Exam]
      parameters:
        - name: accessCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session created with initial state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSessionResponse'
        '404':
          description: Unknown or inactive access code.
        '423':
          description: Exam locked or future-dated.
...
  /api/exams/{sessionId}/submit:
    post:
      summary: Submit exam (voluntary or timeout).
      tags: [Exam]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Submission accepted and objective grading complete.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '409':
          description: Session already closed.
  /api/exams/{sessionId}/grade:
    post:
      summary: Record manual grades for subjective questions.
      tags: [Exam]
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualGradeRequest'
      responses:
        '200':
          description: Manual grades saved and totals recomputed.
...
  /api/exams:
    get:
      summary: List exam sessions (for grading).
      tags: [Exam]
      parameters:
        - in: query
          name: paperId
          schema:
            type: string
        - in: query
          name: userId
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: List of exam sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSessionPage'
  /api/exams/{id}:
    get:
      summary: Get exam session details.
      tags: [Exam]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Exam session details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSessionResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
      description: Zero-based page index.
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size.
...
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    QuestionOption:
      type: object
      properties:
        key:
          type: string
        text:
          type: string
        isCorrect:
          type: boolean
...
    ExamSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        paperVersionId:
          type: string
        userId:
          type: string
        nickname:
          type: string
        username:
          type: string
        score:
          type: integer
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        timeRemainingSeconds:
          type: integer
        autosaveIntervalSeconds:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ExamQuestion'
    ExamQuestion:
      type: object
      properties:
        questionId:
          type: string
        position:
          type: integer
        type:
          type: string
        stem:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOption'
        score:
          type: number
        userAnswer:
          type: string
        awardedScore:
          type: number
        graderNotes:
          type: string
        isCorrect:
          type: boolean
...
    SubmissionResponse:
      type: object
      properties:
        sessionId:
          type: string
        submittedAt:
          type: string
          format: date-time
        objectiveScore:
          type: number
        totalScore:
          type: number
        status:
          type: string
          enum: [SUBMITTED, TIMEOUT]
    ManualGradeRequest:
      type: object
      required: [grades]
      properties:
        grades:
          type: array
          items:
            type: object
            required: [questionId, score]
            properties:
              questionId:
                type: string
              score:
                type: number
                format: float
              notes:
                type: string
...
    ExamSessionPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ExamSessionResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer